+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+## [v1.0.0](https://github.com/pratikghodasara/tp-spotify-pg/compare/0.0.1...1.0.0) - 2024.08.25
+
 ### Added
   - Setting up the Application Features ([#3](https://github.com/pratikghodasara/tp-spotify-pg/pull/3))
     - Added Open, Toggle Visibility and Quit Application features ([f165f5d](https://github.com/pratikghodasara/tp-spotify-pg/commit/f165f5d))